refactor(clawdstrike-policy-event): extract kinds/detection from edr/receipt/mod.rs
