+++ crates/libs/clawdstrike-policy-event/src/edr/receipt/kinds/mod.rs
@@ -1,7 +1,9 @@
+mod detection;
 mod evidence_bundle;
 mod graph_slice;
 mod sensor_state;
 
+pub(crate) use detection::*;
 pub(crate) use evidence_bundle::*;
 pub(crate) use graph_slice::*;
 pub(crate) use sensor_state::*;
